IIO: MPU6150: add option to enable I2C bypass
MPU6150 has two I2C busses. Primary bus is used to communicate with the main
part of the system, the chip acts as slave. Auxiliary bus can be used to
attach third-party sensors which may be accessed through MPU6150 registers.
Auxiliary interface can also be eltctrically connected to primary interface
using built-in analog switches. The patch adds this functionality.

It's also worth to mention that MPU9150 is actually MPU6150 and AK8975 in a
single package, where AK8975 is connected to auxiliary interface. So most
straightforward way to use it is with bypass mode.