fbdev: ft800fb: new driver for FTDI FT800
This commit adds a new framebuffer driver for FTDI FT800 chip.
HW acceleration features aren't used. Framebuffer is kept in the
main RAM and is copied to FT800's internal memory when changes are
detected or continuosly if buffer was mmap-ed and it's not easy to
detect if it has been changed.

The driver also implements input device for FT800 resistive
touchscreen interface.